+++ packages/lwc-compiler/src/rollup-plugins/compat.ts
@@ -1,20 +1,21 @@
 import * as babel from '@babel/core';
 import * as presetCompat from 'babel-preset-compat';
+
 import { BABEL_CONFIG_BASE } from '../babel-plugins';
-import { OutputProxyCompatConfig } from "../compiler/options";
 
-export default function(proxyCompatOption: OutputProxyCompatConfig | undefined) {
-    const config = Object.assign({}, BABEL_CONFIG_BASE, {
-        presets: [
-            [presetCompat, { proxy: true, resolveProxyCompat: proxyCompatOption || { independent: "proxy-compat" } }],
-        ],
-    });
+const BABEL_CONFIG_CONFIG = {
+    ...BABEL_CONFIG_BASE,
+    presets: [
+        [presetCompat, { proxy: true }],
+    ],
+};
 
+export default function() {
     return {
         name: "lwc-compat",
 
         transform(src: string) {
-            const { code, map } = babel.transform(src, config);
+            const { code, map } = babel.transform(src, BABEL_CONFIG_CONFIG);
             return { code, map };
         }
     };