+++ packages/lwc-engine/src/framework/__tests__/watcher.spec.ts
@@ -1,3 +1,4 @@
+import { compileTemplate } from 'test-utils';
 import { createElement, LightningElement } from '../main';
 
 describe('watcher', () => {
@@ -18,10 +19,8 @@ describe('watcher', () => {
 
         it('should rerender the component if any reactive prop changes', () => {
             let counter = 0;
-            function html($api, $cmp) {
-                $cmp.x;
-                return [];
-            }
+
+            const html = compileTemplate(`<template>{x}</template>`);
             class MyComponent2 extends LightningElement {
                 render() {
                     counter++;
@@ -62,26 +61,31 @@ describe('watcher', () => {
 
         it('should rerender the component if any reactive slot changes', () => {
             let counter = 0;
-            function html1($api, $cmp, $slotset, $ctx) {
-                return [$api.s('x', {
-                    key: 0,
-                    attrs: {
-                        name: 'x'
-                    }
-                }, [], $slotset)];
-            }
-            html1.slots = ["x"];
+
+            const childTmpl = compileTemplate(`
+                <template>
+                    <slot name="x"></slot>
+                </template>
+            `);
             class Child extends LightningElement {
                 render() {
                     counter++;
-                    return html1;
+                    return childTmpl;
                 }
             }
-            function html2($api, $cmp) {
-                const r = $cmp.round;
-                return [$api.c('x-child', Child, {}, r === 0 ? [] : [$api.h('p', { key: 0, attrs: { slot: 'x' } }, [])])];
-            }
-            class MyComponent4 extends LightningElement {
+
+            const parentTmpl = compileTemplate(`
+                <template>
+                    <x-child>
+                        <template if:true={round}>
+                            <p slot="x"></p>
+                        </template>
+                    </x-child>
+                </template>
+            `, {
+                modules: { 'x-child': Child }
+            });
+            class Parent extends LightningElement {
                 constructor() {
                     super();
                     this.round = 0;
@@ -90,13 +94,13 @@ describe('watcher', () => {
                     this.round += 1;
                 }
                 render() {
-                    return html2;
+                    return parentTmpl;
                 }
             }
-            MyComponent4.track = { round: 1 };
-            MyComponent4.publicMethods = ['updateRound'];
+            Parent.track = { round: 1 };
+            Parent.publicMethods = ['updateRound'];
 
-            const elm = createElement('x-foo', { is: MyComponent4 });
+            const elm = createElement('x-foo', { is: Parent });
             document.body.appendChild(elm);
             elm.updateRound();
 
@@ -108,10 +112,8 @@ describe('watcher', () => {
         it('should rerender the component if tracked property changes', () => {
             let counter = 0;
             let state;
-            function html($api, $cmp) {
-                $cmp.state.x;
-                return [];
-            }
+
+            const html = compileTemplate(`<template>{state.x}</template>`);
             class MyComponent6 extends LightningElement {
                 state = { x: 0 };
                 constructor() {
@@ -124,6 +126,7 @@ describe('watcher', () => {
                 }
             }
             MyComponent6.track = { state: 1 };
+
             const elm = createElement('x-foo', { is: MyComponent6 });
             document.body.appendChild(elm);
             state.x = 2;
@@ -135,6 +138,7 @@ describe('watcher', () => {
         it('should not rerender the component if a non-reactive state changes', () => {
             let counter = 0;
             let state;
+
             class MyComponent7 extends LightningElement {
                 state = { x: 0 };
                 constructor() {
@@ -145,6 +149,7 @@ describe('watcher', () => {
                     counter++;
                 }
             }
+
             const elm = createElement('x-foo', { is: MyComponent7 });
             document.body.appendChild(elm);
             expect(counter).toBe(1);
@@ -154,60 +159,47 @@ describe('watcher', () => {
             });
         });
 
-        it('should prevent any mutation during the rendering phase', () => {
-            function html($api, $cmp) {
-                $cmp.state.x = 1;
-            }
-            class MyComponent8 extends LightningElement {
-                state = { x: 0 };
-                render() {
-                    return html;
-                }
-            }
-            const elm = createElement('x-foo', { is: MyComponent8 });
-            expect(() => document.body.appendChild(elm)).toThrow();
-        });
-
         it('should compute reactive state per rendering', () => {
             let counter = 0;
-            let state;
-            function html($api, $cmp) {
-                if (counter === 1) {
-                    $cmp.state.x;
-                }
-                return [];
-            }
+
+            const dynamicTmpl = compileTemplate(`<template>{x}</template>`);
+            const staticTmpl = compileTemplate(`<template>static</template>`);
             class MyComponent9 extends LightningElement {
-                state = { x: 0 };
-                constructor() {
-                    super();
-                    state = this.state;
+                x = 0;
+
+                updateTracked() {
+                    this.x++;
                 }
+
                 render() {
                     counter++;
-                    return html;
+                    return counter <= 1 ? dynamicTmpl : staticTmpl;
                 }
             }
-            MyComponent9.track = { state: 1 };
+            MyComponent9.track = { x: 1 };
+            MyComponent9.publicMethods = ['updateTracked'];
+
             const elm = createElement('x-foo', { is: MyComponent9 });
             document.body.appendChild(elm);
             expect(counter).toBe(1);
-            state.x = 1; // this is marked as reactive
+
+            // x is marked as reactive since it's tracked and used in the template
+            elm.updateTracked();
             return Promise.resolve().then(() => {
                 expect(counter).toBe(2);
-                state.x = 2; // this is not longer reactive and should not trigger the rerendering anymore
-                return Promise.resolve().then(() => {
-                    expect(counter).toBe(2);
-                });
+
+                // x is not longer reactive since it's not consumed in the template.
+                // Updating it's value should not trigger the rerendering anymore.
+                elm.updateTracked();
+            }).then(() => {
+                expect(counter).toBe(2);
             });
         });
 
         it('should mark public prop as reactive even if it is used via a getter', () => {
             let counter = 0;
-            function html($api, $cmp) {
-                $cmp.foo;
-                return [];
-            }
+
+            const html = compileTemplate(`<template>{foo}</template>`);
             class MyComponent2 extends LightningElement {
                 get foo() {
                     return this.x;
@@ -223,6 +215,7 @@ describe('watcher', () => {
                 }
             }
             MyComponent2.publicProps = { x: 1 };
+
             const elm = createElement('x-foo', { is: MyComponent2 });
             elm.x = 2;
             document.body.appendChild(elm);
@@ -235,6 +228,7 @@ describe('watcher', () => {
         it('should allow observing public prop via setter', () => {
             let counter = 0;
             let newValue, oldValue;
+
             class MyComponent2 extends LightningElement {
                 set x(value) {
                     counter++;
@@ -246,6 +240,7 @@ describe('watcher', () => {
                 }
             }
             MyComponent2.publicProps = { x: { config: 3 } };
+
             const elm = createElement('x-foo', { is: MyComponent2 });
             elm.x = 2;
             document.body.appendChild(elm);
@@ -259,6 +254,7 @@ describe('watcher', () => {
     describe('#reactivity()', () => {
         it('should react when a reactive array invokes Array.prototype.push()', () => {
             let counter = 0;
+
             class MyComponent1 extends LightningElement {
                 state = { list: [1, 2] };
 
@@ -273,6 +269,7 @@ describe('watcher', () => {
             }
             MyComponent1.track = { state: 1 };
             MyComponent1.publicMethods = ['pushToList'];
+
             const elm = createElement('x-foo', { is: MyComponent1 });
             document.body.appendChild(elm);
             expect(counter).toBe(1);
@@ -283,6 +280,7 @@ describe('watcher', () => {
         });
         it('should react when a reactive array invokes Array.prototype.pop()', () => {
             let counter = 0;
+
             class MyComponent1 extends LightningElement {
                 state = { list: [1, 2] };
 
@@ -297,6 +295,7 @@ describe('watcher', () => {
             }
             MyComponent1.publicMethods = ['popFromList'];
             MyComponent1.track = { state: 1 };
+
             const elm = createElement('x-foo', { is: MyComponent1 });
             document.body.appendChild(elm);
             expect(counter).toBe(1);
@@ -307,6 +306,7 @@ describe('watcher', () => {
         });
         it('should react when a reactive array invokes Array.prototype.unshift()', () => {
             let counter = 0;
+
             class MyComponent1 extends LightningElement {
                 state = { list: [1, 2] };
                 unshiftFromList(value: number) {
@@ -319,6 +319,7 @@ describe('watcher', () => {
             }
             MyComponent1.publicMethods = ['unshiftFromList'];
             MyComponent1.track = { state: 1 };
+
             const elm = createElement('x-foo', { is: MyComponent1 });
             document.body.appendChild(elm);
             expect(counter).toBe(1);