+++ packages/lwc-jest-transformer/src/test/modules/example/contentAssetUrl/__tests__/contentAssetUrl.test.js
@@ -1,0 +1,34 @@
+import { createElement } from 'lwc';
+import ContentAssetUrl from 'example/contentAssetUrl';
+
+jest.mock('@salesforce/contentAssetUrl/mocked', () => {
+    return { default: "value set in test"};
+}, { virtual: true });
+
+afterEach(() => {
+    while (document.body.firstChild) {
+        document.body.removeChild(document.body.firstChild);
+    }
+});
+
+describe('example-content-asset-url', () => {
+    it('default snapshot', () => {
+        const element = createElement('example-content-asset-url', { is: ContentAssetUrl });
+        document.body.appendChild(element);
+        expect(element).toMatchSnapshot();
+    });
+
+    it('returns default value as import path', () => {
+        const element = createElement('example-content-asset-url', { is: ContentAssetUrl });
+        document.body.appendChild(element);
+        const resourceUrl = document.body.querySelector(".unmockedAsset").textContent;
+        expect(resourceUrl).toBe("unmocked");
+    });
+
+    it('returns value from mock defined in test file', () => {
+        const element = createElement('example-content-asset-url', { is: ContentAssetUrl });
+        document.body.appendChild(element);
+        const resourceUrl = document.body.querySelector(".mockedAsset").textContent;
+        expect(resourceUrl).toBe("value set in test");
+    });
+});